% Generated by roxygen2 (4.0.2): do not edit by hand
\name{adonis.comp}
\alias{adonis.comp}
\title{Compute single F score}
\usage{
adonis.comp(dis, groups, permutations = 99, f.perms = FALSE,
  svQTL = FALSE, approx = TRUE)
}
\arguments{
\item{dis}{the distance object of the transcript relative expression.}

\item{groups}{a factor with the group information.}

\item{permutations}{the number of permutations.}

\item{f.perms}{should the permuted F scores be returned instead of the
real F score. Default is FALSE.}

\item{svQTL}{should svQTL test be performed instead of sQTL. Default is FALSE.}

\item{approx}{should the asymptotic distribution be used instead of permutations.
Default is TRUE.}
}
\value{
a vector with the F score or the permuted F scores.
}
\description{
Compute F score using the \code{vegan} package.
}
\author{
Jean Monlong
}
\keyword{internal}

