% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.events.R
\name{classify.events}
\alias{classify.events}
\title{Classify splicing events}
\usage{
classify.events(df, trans.struct)
}
\arguments{
\item{df}{a data.frame that includes pairs of transcript IDs in columns 'transId1' and 'transId2'.}

\item{trans.struct}{a data.frame with the transcript structure, i.e. the location of its exons and eventually its UTRs. See Details for format.}
}
\value{
a list with
\item{res}{input data.frame with two new columns: 'classCode' and 'classEvent'. See Details for interpretation.}
\item{stats}{a data.frame with the occurence of each event in the data.}
}
\description{
Find the splicing event(s) that differenciate pairs of transcripts.
}
\details{
The transcript structure...

The classification code follows mostly the one defined by AStalavista (\url{http://genome.crg.es/astalavista/FAQ.html})....
}
\author{
Jean Monlong
}

