% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compFscore}
\alias{compFscore}
\title{F score computation}
\usage{
compFscore(geno.df, tre.dist, tre.df, svQTL = FALSE)
}
\arguments{
\item{geno.df}{a data.frame of one row with the genotype information for each sample.}

\item{tre.dist}{a distance object from the transcript relative expression.}

\item{tre.df}{a data.frame with the transcript relative expression.}

\item{svQTL}{should svQTL test be performed in addition to sQTL. Default is FALSE.}
}
\value{
a data.frame with columns:
\item{F}{the F score.}
\item{nb.groups}{the number of groups created by the genotypes.}
\item{md}{the maximum difference in splicing ratios between genotype groups.}
\item{tr.first, tr.second}{the two transcripts that change the most.}
}
\description{
Compute the F score, max diff ratio difference.
}
\author{
Jean Monlong
}
\keyword{internal}

