% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compute.null}
\alias{compute.null}
\title{F score null distribution}
\usage{
compute.null(F.lead, dis, nb.gp, min.nb.ext.scores = 1000,
  nb.perm.max = 1e+06, svQTL = FALSE, approx = TRUE)
}
\arguments{
\item{F.lead}{the highest F score to drive the permutation optimization.}

\item{dis}{the distance object of the transcript relative expression.}

\item{nb.gp}{the number of group to simulate.}

\item{min.nb.ext.scores}{the minimum number of permuted score higher than
'F.lead' to allow the computation to stop. Default is 1000.}

\item{nb.perm.max}{the maximum number of permutations. Default is 1e6.}

\item{svQTL}{should svQTL test be performed instead of sQTL. Default is FALSE.}

\item{approx}{should the asymptotic distribution be used instead of permutations.
Default is TRUE.}
}
\value{
a list with
\item{FP}{a vector of the permuted F scores.}
\item{nbP.tot}{the number of permuted F scores.}
}
\description{
Compute the null distribution for the F scores.
}
\author{
Jean Monlong
}
\keyword{internal}

