% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hellingerDist}
\alias{hellingerDist}
\title{Hellinger distance computation}
\usage{
hellingerDist(sr)
}
\arguments{
\item{sr}{a matrix or data.frame with the transcript relative
expression for each sample (column).}
}
\value{
a matrix with the pairwise distance between samples.
}
\description{
Computes Hellinger distance for each pair of columns from the input
data.
}
\author{
Jean Monlong
}
\keyword{internal}

