% Generated by roxygen2 (4.0.2): do not edit by hand
\name{index.genotype}
\alias{index.genotype}
\title{Index genotype file}
\usage{
index.genotype(file)
}
\arguments{
\item{file}{the name of the genotype file}
}
\value{
the name of the indexed genotype file.
}
\description{
Index genotype file. The file must be ordered by chr, start, end position. Moreover,
the first three columns must be 'chr', 'start' and 'end' information.
}
\author{
Jean Monlong
}

