% Generated by roxygen2 (4.0.2): do not edit by hand
\name{md.trans}
\alias{md.trans}
\title{MaxDiff splicing ratios computation}
\usage{
md.trans(sr.o, groups.o, samples)
}
\arguments{
\item{sr.o}{a matrix or data.frame with the transcript relative expression.}

\item{groups.o}{a factor with the genotype groups for each sample.}

\item{samples}{a vector with the names of samples used in the genotype
groups information.}
}
\value{
a list with:
\item{md}{the maximum difference in splicing ratios between genotype groups.}
\item{tr.first, tr.second}{the two transcripts that change the most.}
}
\description{
Compute the maximum difference in transcript relative expression
between genotype groups.
}
\author{
Jean Monlong
}
\keyword{internal}

