% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nbDiffPt}
\alias{nbDiffPt}
\title{Number of different splicing ratio points}
\usage{
nbDiffPt(sr)
}
\arguments{
\item{sr}{a data.frame with the splicing ratios (transcript x sample).}
}
\value{
the number of different splicing ratios
}
\description{
Compute the number of different splicing ratios. Too few splicing ratios is
a concern for the permutation process. Indeed if many samples exactly the same
splicing ratios, the permutation might create many identical permuted F scores.
}
\author{
Jean Monlong
}
\keyword{internal}

