% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutest.betadisper.R
\name{permutest.betadisper}
\alias{permutest.betadisper}
\title{Betadisper permutation hack}
\usage{
permutest.betadisper(x, pairwise = FALSE, control = permute::how(nperm =
  999), ...)
}
\arguments{
\item{x}{an object of class 'betadisper' the result of a call to 'betadisper'.}

\item{pairwise}{logical; perform pairwise comparisons of group means ?}

\item{control}{a list of control values for the permutations as returned by
the function 'how'.}

\item{...}{Arguments passed to other methods.}
}
\value{
a list of class 'permutest.betadisper'
}
\description{
The older version of 'permutest.betadisper' which returned the permuted F scores.
}
\author{
Jean Monlong
}
\keyword{internal}

