% Generated by roxygen2 (4.0.2): do not edit by hand
\name{relativize}
\alias{relativize}
\title{Relative expression computation}
\usage{
relativize(x, min.gene.exp = 0.01)
}
\arguments{
\item{x}{a vector with of transcript expression}

\item{min.gene.exp}{the minimum gene expression in the sample. If the
gene expression if too low it is usually safer to remove from the
analysis.}
}
\value{
a vector with the relative expression.
}
\description{
Compute the relative expression of transcripts.
}
\author{
Jean Monlong
}
\keyword{internal}

