% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sqtls}
\alias{sqtls}
\title{sQTLs retrieval}
\usage{
sqtls(res.df, FDR = 0.01, md.min = 0.01, out.pdf = NULL,
  svQTL.removal = TRUE)
}
\arguments{
\item{res.df}{a data.frame with the results from 'sqtl.seeker' with the P-values
for each gene/SNP test.}

\item{FDR}{the False Discovery Rate to call an association significant.}

\item{md.min}{the minimum MD (Maximum Difference) in relative expression desired. Maximum difference in relative expression (MD) gives an idea of the effect size of the association.}

\item{out.pdf}{the name of the pdf file to create. If NULL (default), no pdf output
will be created. If non-NULL, the distribution of the P-values and a semi-volcano plot
(P-value vs MD) will be shown.}

\item{svQTL.removal}{if TRUE (and column 'pv.svQTL' is present in 'res.df') significant
sQTL which are also significant svQTLs are not reported.}
}
\value{
a subset of the input data.frame with only significant sQTLs.
}
\description{
Retrieves sQTLs after multiple test correction and, potentially removing svQTLs.
}
\author{
Jean Monlong
}

