\name{sQTLseekeR-package}
\alias{sQTLseekeR-package}
\alias{sQTLseekeR}
\docType{package}
\title{Splicing QTL discovery tool}
\description{Association study between genotypic information and transcript relative expression from RNASeq, using a distance-based non-parametric MANOVA approach.}
\details{
\tabular{ll}{
Package: \tab sQTLseekeR\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2014-10-01\cr
License: \tab GPL-2\cr
}
}
\author{Jean Monlong, Miquel Calvo, Roderic Guigo.

Maintainer: Jean Monlong <jean.monlong@gmail.com>}
\references{Monlong, J. et al. Identification of genetic variants associated with alternative splicing using sQTLseekeR. Nat. Commun. 5:4698 doi: 10.1038/ncomms5698 (2014).}
\keyword{splicing, qtl, rna-seq, sequencing, transcriptome}
\examples{
}
