% Generated by roxygen2 (4.0.2): do not edit by hand
\name{te.dispersion}
\alias{te.dispersion}
\title{Dispersion computation}
\usage{
te.dispersion(d)
}
\arguments{
\item{d}{a distance matrix with pairwise distance between the samples.}
}
\value{
a value for the dispersion
}
\description{
Compute the dispersion from a distance matrix. The dispersion is computed as
the average distance to the centroid.
}
\author{
Jean Monlong
}
\keyword{internal}

