% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prepare.trans.exp}
\alias{prepare.trans.exp}
\title{Transcript expression preparation}
\usage{
prepare.trans.exp(te.df, min.transcript.exp = 0.01, min.gene.exp = 0.01,
  min.dispersion = 0.1)
}
\arguments{
\item{te.df}{a data.frame with the transcript expression. The first
two columns are the transcript and gene ids, named 'trId' and
'geneId', then the values for each sample the other columns.}

\item{min.transcript.exp}{the minimum transcript expression. Transcript
with lower expression in all samples are removed.}

\item{min.gene.exp}{the minimum gene expression. Samples with lower
gene expression are removed from the analysis of this gene.}

\item{min.dispersion}{the minimum dispersion of the transcript
relative expression. Genes with lower dispersion are removed.}
}
\value{
a data.frame with the relative transcript expression for the
genes to study.
}
\description{
Prepare the transcript expression matrix. Transcript with low
expression are removed. Then gene expressing only one transcript
are removed because not informative for splicing analysis. Finally
the relative expression of the transcripts is retrieved and genes with
low dispersion are removed. Also genes with just a few (<25)
different splicing patterns are removed as they don't fit the
permutation process.
}
\details{
After removal of inappropriate transcripts/genes, transcript
relative expression is computed by dividing its expression
by the total expression of all transcripts for the gene.
}
\author{
Jean Monlong
}

