% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sqtl.seeker}
\alias{sqtl.seeker}
\title{sQTL seeker}
\usage{
sqtl.seeker(tre.df, genotype.f, gene.loc, genic.window = 5000,
  min.nb.ext.scores = 1000, nb.perm.max = 1e+06,
  nb.perm.max.svQTL = 1e+05, svQTL = FALSE, approx = TRUE)
}
\arguments{
\item{tre.df}{a data.frame with transcript relative expression
produced by 'prepare.trans.exp'.}

\item{genotype.f}{the name of the genotype file. This file need to
be ordered by position, compressed and indexed using tabix (samtools).
Must have column 'snpId'.}

\item{gene.loc}{a data.frame with the genes location. Columns 'chr', 'start',
'end' and 'geneId' are required.}

\item{genic.window}{the window(bp) around the gene in which the SNPs are tested. Default is 5000 (i.e. 5kb).}

\item{min.nb.ext.scores}{the minimum number of permuted score higher than
'F.lead' to allow the computation to stop. Default is 1000.}

\item{nb.perm.max}{the maximum number of permutations. Default is 1e6.}

\item{nb.perm.max.svQTL}{the maximum number of permutations for the svQTL computation. Default is 1e5.}

\item{svQTL}{should svQTLs test be performed in addition to sQTLs. Default is FALSE. Warning:
computation of svQTLs cannot rely on asymptotic approximation, hence the heavy permutations will
considerably increase the running time.}

\item{approx}{should the asymptotic distribution be used instead of permutations.
Default is TRUE.}
}
\value{
a data.frame with columns
\item{snpId}{the SNP name}
\item{geneId}{the gene name}
\item{F}{the F score}
\item{nb.groups}{the number of genotype groups}
\item{pv}{the P-value}
\item{nb.perms}{the number of permutation used for the P-value computation}
}
\description{
\code{sqtl.seeker} is the main function of \code{sQTLseekeR} package. From transcript relative expression,
prepared using \code{prepare.trans.exp}, information about the gene location and the path to an ordered genotype
file, indexed by function \code{index.genotype}, potential association between each SNP and the transcript
relative expression is tested. Potentially, svQTL, i.e. SNPs affecting splicing variability can also be tested
to pinpoint potential false sQTL (see Details).
}
\details{
A set of filters is automatically used to remove SNPs which are unpractical or not informative. Precisely, these
filters remove SNP :
\itemize{
\item{with more than 3 unknown genotypes}
\item{with less than 5 samples in any genotype group}
\item{with less than 5 different splicing pattern (needed for permutation efficiency) in any genotype group}}

Testing difference in transcript relative expression between genotype groups assumes homogeneity of the variances
in these groups. Testing this assumption is more complex and computationnally intensive but if needed the user
can choose to test for svQTL, i.e. gene/SNPs where this assumption is violated, by using the \code{svQTL=TRUE}. This
test is run in parallel to the sQTL tests, but the copmutation time will be considerably higher. For this reason, another
parameter can be tweaked, \code{nb.perm.max.svQTL}, to reduce the number of permutation for the svQTL tests if needed for
feasibility reasons.

DETAILS ON APPROXIMATION

DETAILS ON MD OUTPUT
}
\author{
Jean Monlong
}

