% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sqtl.seeker}
\alias{sqtl.seeker}
\title{sQTL seeker}
\usage{
sqtl.seeker(tre.df, genotype.f, gene.loc, genic.window = 5000,
  min.nb.ext.scores = 1000, nb.perm.max = 1e+06, svQTL = FALSE,
  approx = TRUE)
}
\arguments{
\item{tre.df}{a data.frame with transcript relative expression
produced by 'prepare.trans.exp'.}

\item{genotype.f}{the name of the genotype file. This file need to
be ordered by position, compressed and indexed using tabix (samtools).
Must have column 'snpId'.}

\item{gene.loc}{a data.frame with the genes location. Columns 'chr', 'start',
'end' and 'geneId' are required.}

\item{genic.window}{the window(bp) around the gene in which the SNPs are tested. Default is 5000 (i.e. 5kb).}

\item{min.nb.ext.scores}{the minimum number of permuted score higher than
'F.lead' to allow the computation to stop. Default is 1000.}

\item{nb.perm.max}{the maximum number of permutations. Default is 1e6.}

\item{svQTL}{should svQTL test be performed instead of sQTL. Default is FALSE.}

\item{approx}{should the asymptotic distribution be used instead of permutations.
Default is TRUE.}
}
\value{
a data.frame with columns
\item{snpId}{the SNP name}
\item{geneId}{the gene name}
\item{F}{the F score}
\item{nb.groups}{the number of genotype groups}
\item{pv}{the P-value}
\item{nb.perms}{the number of permutation used for the P-value computation}
}
\description{
Run sQTLseekeR on a set of genes.
}
\author{
Jean Monlong
}

